pyproject.toml
[project]
name = "gh-project-man"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "rich>=14.0.0",
]
